backend/requirements.txt
# Backend requirements for unified medical system

# Django framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-environ==0.11.2
django-filter==23.3

# Database drivers
psycopg2-binary==2.9.7
pymongo==4.6.0
mongoengine==0.27.0
redis==5.0.1

# Async task processing
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Speech-to-Text engines
openai-whisper==20231117
azure-cognitiveservices-speech==1.34.0
speechrecognition==3.10.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# NLP and ML
numpy==1.24.3
torch==2.1.1
transformers==4.36.0
huggingface-hub==0.19.4
accelerate==0.25.0
datasets==2.15.0
sentence-transformers==2.2.2

# NVIDIA NIM LLM API
openai==1.12.0

# OpenAI Whisper classic (stable)
openai-whisper==20231117

# Groq backup LLM
# groq==0.4.1

# Real-time Whisper transcription (commented out for stability)
# faster-whisper==0.10.0
# pyaudio==0.2.11
# webrtcvad==2.0.10

# PDF generation
reportlab==4.0.7
Pillow==10.1.0
weasyprint==60.2

# WebSocket support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Authentication and security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Data validation and serialization
pydantic==2.5.0
marshmallow==3.20.1
jsonschema==4.20.0

# Utilities
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Monitoring and logging
sentry-sdk==1.38.0
structlog==23.2.0
django-debug-toolbar==4.2.0

# CLI utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1